Renamed autosave.cpp/h to backupservice.cpp/h and moved the cpp file to the correct directory.
