+++ src/ui/frmpreferences.cpp
@@ -6,7 +6,7 @@
 #include "include/Extensions/extensionsloader.h"
 #include "include/notepadqq.h"
 #include "include/keygrabber.h"
-#include "include/Sessions/autosave.h"
+#include "include/Sessions/backupservice.h"
 #include "include/stats.h"
 #include <QFileDialog>
 #include <QSortFilterProxyModel>
